refactor(lib): promote formatValue out of status analytics internals
DatabaseOverview (databases feature) needed the shared byte formatting and
was reaching into status/analytics internals — first via the deleted
formatBytes (broke stage's type-check after the theme PR), then via the
analytics formatValue path in the hotfix. The formatter is app-generic, so
it moves to src/lib with a chart-agnostic ValueFormatter union; analytics'
AxisFormatter becomes an alias and every consumer imports from the lib.
