+++ src/features/instance/status/analytics/primitives/StackedAreaChart.tsx
@@ -1,10 +1,10 @@
+import { formatValue } from '@/lib/formatValue';
 import { useMemo } from 'react';
 import { Area, AreaChart, CartesianGrid, Legend, Line, ResponsiveContainer, Tooltip, XAxis, YAxis } from 'recharts';
 import { NODE_PALETTE } from '../lib/nodeColors.ts';
 import { getChartColors, useResolvedTheme } from '../lib/theme.ts';
 import { formatAxisTick, formatTooltipTime } from '../lib/time.ts';
 import type { AxisFormatter, AxisSpec, SeriesData } from '../types/analytics.ts';
-import { formatValue } from './formatValue.ts';
 import { sortByMagnitude } from './sortByMagnitude.ts';
 import { tooltipContentStyle, tooltipLabelStyle } from './tooltipStyle.ts';
